Answer the review: one layer owner, one plan, one top row, and a gate that reads Swift
One VideoSurfaceView owns a source's layer slot: the column skips the
promoted tile, updates re-home a changed source, and teardown releases
the slot only if it still holds it. The dead instrument-rack toggle
leaves the menu. Nil survives the section selection round trip for
collapsed presentations. Every planner face edits ONE MissionPlanModel
and the pill's summary follows it. The columns toggle and the FCU
share one top row and cannot collide. The structure gate now measures
Swift file lengths — it caught the planner file the same hour it
learned to look — and the app and planner files are split under the
ceiling. Command intent families get a named constant; the flow layout
wraps against its real bounds and centers each row; clearing typed
text no longer deletes a chip; compound procedure labels admit the
space; a write-only width and a doubled height writer are gone.
