+++ clients/apple/App/SituationMenuView.swift
@@ -26,8 +26,7 @@ struct SituationMenuView: View {
         NavigationStack {
             List {
                 receptionSection
-                instrumentsSection
-                flightsSection
+                    flightsSection
                 if let message = model.errorMessage {
                     Section("Problems") {
                         Text(message).font(.footnote)
@@ -42,23 +41,6 @@ struct SituationMenuView: View {
                 }
             }
         }
-    }
-
-    /// The rack shares the screen with the map; this is its switch.
-    private var instrumentsSection: some View {
-        Section {
-            Toggle("Instrument rack", isOn: rackPresented)
-        } footer: {
-            Text("A vertical stack of live instruments beside the map. "
-                + "Pick the profile and the host connection on the rack itself.")
-        }
-    }
-
-    private var rackPresented: Binding<Bool> {
-        Binding(
-            get: { UserDefaults.standard.bool(forKey: "pilotageRackPresented") },
-            set: { UserDefaults.standard.set($0, forKey: "pilotageRackPresented") }
-        )
     }
 
     /// What the radios are doing, said before it is explained.
