+++ clients/apple/App/VideoFrameHub.swift
@@ -40,10 +40,14 @@ final class VideoFrameHub: @unchecked Sendable {
         layer.contents = image?.cgImage
     }
 
-    /// Unbinds whatever layer serves this source.
-    func detach(source: UInt8) {
+    /// Unbinds a layer from a source — but only the layer that owns
+    /// the slot. A stale view tearing down must never evict the live
+    /// one that replaced it.
+    func detach(layer: CALayer, source: UInt8) {
         lock.lock()
-        layers.removeValue(forKey: source)
+        if layers[source] === layer {
+            layers.removeValue(forKey: source)
+        }
         lock.unlock()
     }
 
@@ -60,8 +64,6 @@ final class VideoFrameHub: @unchecked Sendable {
         }
     }
 }
-
-
 /// Hosts one video source's picture in a plain layer the frame hub
 /// paints directly: no picture ever crosses SwiftUI, so a sixty-hertz
 /// feed re-evaluates nothing but its own layer contents.
@@ -69,14 +71,36 @@ struct VideoSurfaceView: UIViewRepresentable {
     let hub: VideoFrameHub
     let source: UInt8
 
+    final class Coordinator {
+        var hub: VideoFrameHub?
+        var source: UInt8 = 0
+    }
+
+    func makeCoordinator() -> Coordinator {
+        Coordinator()
+    }
+
     func makeUIView(context: Context) -> UIView {
         let view = UIView()
         view.layer.contentsGravity = .resizeAspect
+        context.coordinator.hub = hub
+        context.coordinator.source = source
         hub.attach(layer: view.layer, source: source)
         return view
     }
 
     func updateUIView(_ view: UIView, context: Context) {
+        if context.coordinator.source != source {
+            context.coordinator.hub?.detach(layer: view.layer, source: context.coordinator.source)
+        }
+        context.coordinator.hub = hub
+        context.coordinator.source = source
         hub.attach(layer: view.layer, source: source)
+    }
+
+    static func dismantleUIView(_ uiView: UIView, coordinator: Coordinator) {
+        // A departing tile releases its slot, so the hub never paints
+        // into a layer that left the hierarchy.
+        coordinator.hub?.detach(layer: uiView.layer, source: coordinator.source)
     }
 }
